correct annotation

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -4,7 +4,7 @@
 
 from typing import List, Union
 
-MixedList = List[Union[float, int]]
+MixedList = List[Union[int, float]]
 
 
 def sum_mixed_list(mxd_lst: MixedList) -> float:
